Add SSL integration tests and support for SSL connections in ActiveMQ (#30)
* Add SSL integration tests and support for SSL connections in ActiveMQ

- Introduced OpenwireMultiConnectionTest for testing multiple simultaneous connections to different brokers.
- Created main integration benchmark test runner in main.cpp.
- Updated CMakeLists.txt to include new integration benchmark tests.
- Added OpenwireSslConnectionTest and OpenwireSslSimpleTest for SSL connection testing.
- Enhanced IntegrationCommon to provide SSL OpenWire URL.
- Updated docker-compose.yml to include SSL broker configuration.
- Added SSL broker configuration in activemq-ssl.xml.
- Created script to generate self-signed certificates for SSL testing.


* Add comprehensive SSL integration tests for OpenWire protocol


* Replace README.txt with README.md for improved documentation format and content


* Update SSL configuration in docker-compose and activemq-ssl.xml for clarity and port remapping


* Enhance OpenSSLSocket to handle non-blocking sockets and improve error logging during SSL connection and handshake


* Enhance Openwire SSL Tests with Improved Listener and Transaction Handling

- Refactored OpenwireSslSlowListenerTest to include a dedicated SlowListener class that tracks message counts and thread IDs.
- Updated testSlowListener to validate that all messages are processed by a single thread and to ensure message counts are accurate.
- Expanded OpenwireSslTransactionTest with additional transaction scenarios, including batch sending and rollback handling.
- Improved testSendRollbackCommitRollback to ensure messages are correctly rolled back and re-sent.
- Added testSendSessionClose to verify behavior when closing sessions during message transactions.
- Introduced testWithTTLSet to validate message expiration handling in transactions.
- Enhanced OpenwireSslVirtualTopicTest to streamline virtual topic message sending and receiving with various acknowledgment modes.


* Implement automatic SSL certificate generation and update documentation


* Refactor integration tests to separate OpenWire and STOMP tests, and update documentation for clarity


* Enhance SSL integration tests with comprehensive coverage and add new test cases for OpenWire SSL features


* Refactor OpenWire SSL Tests and Enhance SSL Configuration

- Updated OpenwireSslSlowListenerTest to improve namespace formatting.
- Refactored OpenwireSslTempDestinationTest to replace inheritance from TempDestinationTest with CMSTestFixture, adding new test cases for message handling and connection behavior.
- Enhanced OpenwireSslTransactionTest and OpenwireSslXATransactionsTest to utilize Google Test framework, improving structure and readability.
- Updated IntegrationCommon to include SSL peer verification details.
- Modified CMake presets to enable SSL support.
- Adjusted docker-compose and SSL configuration files to reflect new paths for ActiveMQ and certificate files.


* Add <cstdint> include to SocketFileDescriptor and FileDescriptor headers


* Update CI configuration for Windows SSL integration tests and refine Docker Compose comments


* Remove commented SSL configuration examples from main.cpp


* Refactor SSL support configuration to use AMQCPP_USE_SSL instead of HAVE_OPENSSL


* Remove Windows SSL integration test job from CI configuration


* Refactor thread destruction logic to improve synchronization and prevent early returns


* Refactor SSL test source inclusion in CMakeLists.txt for transport and decaf modules


* Update integration test commands in CI configuration to use exact label matching


* Add test suite delimiter configuration for CTest in settings.json


* Refactor thread destruction logic to improve synchronization and prevent early returns


* Refactor CmsSendWithAsyncCallbackTest to improve connection and destination handling


* Disable TLS 1.3 in OpenSSLContextSpi to prevent race conditions during concurrent SSL_read and SSL_write operations


* Add crash handlers for Flight Recorder to capture dumps on exceptions and signals


---------

+++ activemq-cpp/src/main/decaf/internal/net/SocketFileDescriptor.cpp
@@ -16,6 +16,7 @@
  */
 
 #include "SocketFileDescriptor.h"
+#include <cstdint>
 
 using namespace decaf;
 using namespace decaf::io;
@@ -23,7 +24,7 @@ using namespace decaf::internal;
 using namespace decaf::internal::net;
 
 ////////////////////////////////////////////////////////////////////////////////
-SocketFileDescriptor::SocketFileDescriptor( long value ) : FileDescriptor( value, false ) {
+SocketFileDescriptor::SocketFileDescriptor( intptr_t value ) : FileDescriptor( value, false ) {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -31,6 +32,6 @@ SocketFileDescriptor::~SocketFileDescriptor() {
 }
 
 ////////////////////////////////////////////////////////////////////////////////
-long SocketFileDescriptor::getValue() const {
+intptr_t SocketFileDescriptor::getValue() const {
     return this->descriptor;
 }
