+++ activemq-cpp/src/main/decaf/internal/net/SocketFileDescriptor.h
@@ -18,6 +18,7 @@
 #ifndef _DECAF_INTERNAL_NET_SOCKETFILEDESCRIPTOR_H_
 #define _DECAF_INTERNAL_NET_SOCKETFILEDESCRIPTOR_H_
 
+#include <cstdint>
 #include <decaf/util/Config.h>
 
 #include <decaf/io/FileDescriptor.h>
@@ -34,7 +35,7 @@ namespace net {
     class DECAF_API SocketFileDescriptor : public decaf::io::FileDescriptor {
     public:
 
-        SocketFileDescriptor( long value );
+        SocketFileDescriptor( intptr_t value );
 
         virtual ~SocketFileDescriptor();
 
@@ -43,7 +44,7 @@ namespace net {
          *
          * @return a FileDescriptor value.
          */
-        long getValue() const;
+        intptr_t getValue() const;
 
     };
